% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arthurs_cell_annotation_functions.R
\name{celltype_annotation_by_cluster}
\alias{celltype_annotation_by_cluster}
\title{match scRNA clusters with bulk annotation}
\usage{
celltype_annotation_by_cluster(dataset, min_percent = 0.5)
}
\arguments{
\item{dataset}{dataset object output by metadata_reprocessing}

\item{min_percent}{minimun percentage for a cell cluster being annotated as certain cell type}
}
\value{
an data_set object
}
\description{
label each scRNA cluster in Exp_Seurat@ident as a bulk cell type, if more than min_percent of the
cells from the cluster has certain bulk cell label.
}
