% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arthurs_cell_annotation_functions.R
\name{data_preprocessing}
\alias{data_preprocessing}
\title{process the list of bulk samples, the single cell matrix into a single object for downstream analysis}
\usage{
data_preprocessing(bulk_list, sc_mat, gene_id = NULL)
}
\arguments{
\item{bulk_list}{input bulk list}

\item{sc_mat}{input single cell / super cell matrix}

\item{gene_id}{input gene id vector}
}
\value{
an data_set object
}
\description{
This function loads the bulk_list object output by "bulk_list_normalize", a single cell matrix and a gene id vector
each row of the single cell matrix is a row, and each column is a gene. 
The gene id vector must be identical to the gene order of the single cell matrix, and the gene order of bulk_list
}
