% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arthurs_cell_annotation_functions.R
\name{metadata_reprocessing}
\alias{metadata_reprocessing}
\title{match bulk annotation with scRNA groups}
\usage{
metadata_reprocessing(dataset, Exp_Seurat, relavant_cells,
  filtering_based_on_category = NULL, filtering_threshold = 0.01)
}
\arguments{
\item{dataset}{dataset object output by secondary_cor}

\item{Exp_Seurat}{Seurat object with @metadata}

\item{relavant_cells}{A vector cell position id indicating the cells that's relveant for the analysis.}

\item{filtering_based_on_category}{The type of category that filtering will happen on,must be present in Exp_Seurat@meta.data}
}
\value{
an data_set object
}
\description{
For a given meta data label, this function calculate with annotated cell type present in the meta data label, 
and which meta data label is irrelavant for the current dataset. The filtering process ensure that only ident
group that present in at least one of the category will be kept. presence of a group means that more than
filtering_threshold % of a category containing cells from given group. filtering_threshold=NULL means no filtering
}
