% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arthurs_cell_annotation_functions.R
\name{plot_bulk_bulk_cor}
\alias{plot_bulk_bulk_cor}
\title{plot the correlation between all bulk samples}
\usage{
plot_bulk_bulk_cor(dataset)
}
\arguments{
\item{dataset}{dataset object ouput by data_preprocessing}
}
\description{
plot the correlation between all bulk samples
}
