% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arthurs_cell_annotation_functions.R
\name{plot_group_annotation_TSNE}
\alias{plot_group_annotation_TSNE}
\title{plot the group annotation on TSNE}
\usage{
plot_group_annotation_TSNE(dataset, Exp_Seurat, plot_sets = NULL)
}
\arguments{
\item{dataset}{dataset object ouput by secondary_cor}

\item{Exp_Seurat}{Seurat object with @dr$tsne slot calcualted}

\item{plot_sets}{which bulk sample set to plot}
}
\description{
plot the group annotation on TSNE
}
