% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arthurs_cell_annotation_functions.R
\name{plot_SC_sampleSet_cor_2nd_TSNE}
\alias{plot_SC_sampleSet_cor_2nd_TSNE}
\title{plot the 2nd correlation between sc and bulk sample on TSNE}
\usage{
plot_SC_sampleSet_cor_2nd_TSNE(dataset, Exp_Seurat, color_contrast = 1,
  plot_sets = NULL)
}
\arguments{
\item{dataset}{dataset object ouput by secondary_cor}

\item{Exp_Seurat}{Seurat object with @dr$tsne slot calcualted}

\item{color_contrast}{positive number for adjusting color contrast, default=1}

\item{plot_sets}{which bulk sample set to plot}
}
\description{
plot the 2nd correlation between sc and bulk sample on TSNE
}
