% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arthurs_cell_annotation_functions.R
\name{plot_SC_sampleSet_cor_distribution}
\alias{plot_SC_sampleSet_cor_distribution}
\title{plot the distribution of sc to mean expresion of each sample set}
\usage{
plot_SC_sampleSet_cor_distribution(dataset, plot_sets = NULL)
}
\arguments{
\item{dataset}{dataset object ouput by data_preprocessing}

\item{plot_sets}{which plot set to pick (number_id or name)}
}
\description{
plot the distribution of sc to mean expresion of each sample set
}
