% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arthurs_cell_annotation_functions.R
\name{plot_SC_sampleSet_cor_TSNE}
\alias{plot_SC_sampleSet_cor_TSNE}
\title{plot the correlation between sc and bulk set mean on TSNE}
\usage{
plot_SC_sampleSet_cor_TSNE(dataset, Exp_Seurat, thresholds = NULL,
  color_contrast = 1, plot_sets = NULL)
}
\arguments{
\item{dataset}{dataset object ouput by data_preprocessing}

\item{thresholds}{threshold for picking the highly similar cells, cells have similarity lower than threshold will be mark gray}
}
\description{
plot the correlation between sc and bulk set mean on TSNE
}
