% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arthurs_cell_annotation_functions.R
\name{plot_scMeta}
\alias{plot_scMeta}
\title{plot the group annotation one by one on TSNE}
\usage{
plot_scMeta(dataset, Exp_Seurat, tagcol, main = NULL, addtext = T)
}
\arguments{
\item{dataset}{dataset object with sc_mat slot}

\item{Exp_Seurat}{Seurat object with @dr$tsne slot and @meta.data}

\item{tagcol}{which column of Exp_Seurat@meta.data shoule we plot (column name)}
}
\description{
plot the group annotation one by one on TSNE
}
