% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arthurs_cell_annotation_functions.R
\name{secondary_cor}
\alias{secondary_cor}
\title{calculate the 2nd correlation}
\usage{
secondary_cor(dataset, threshold)
}
\arguments{
\item{dataset}{dataset object output by data_preprocessing}

\item{threshold}{vector of values that between 0~1}
}
\value{
an data_set object
}
\description{
This function calculate the 2nd correlation, basically force cells that passed threshold to pick side within a bulk set.
}
